% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkly_format.R
\name{darkly_format}
\alias{darkly_format}
\title{Format helper function for labels}
\usage{
darkly_format(x, labelx, n.decimals = 2)
}
\arguments{
\item{x}{input vector frame requires input}

\item{labelx}{label identifier, takes "round" or "percent', all others default to character}

\item{n.decimals}{when labelx=round digits parameter for round(), when labelx=percent accuracy parameter for scales::percent()}
}
\description{
Format helper function for labels
}
