% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darklyplot.R
\name{darklyplot}
\alias{darklyplot}
\title{Darkly plot function}
\usage{
darklyplot(
  df = indata,
  column,
  col = "white",
  n.decimals = 0,
  refline = FALSE,
  refValue = 0,
  refCol = NA,
  shade = FALSE,
  minCol = "#6CB23F",
  maxCol = "#FD5305",
  firstCol = "white",
  lastCol = "#00AFEF",
  shadeCol = "dodgerblue",
  shadeAlpha = 0.15,
  labelx = "round"
)
}
\arguments{
\item{df}{input data frame}

\item{column}{input data column name to plot}
}
\description{
A dark themed plot
}
\keyword{theme}
