% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_rate.R
\docType{data}
\name{mtg_rate}
\alias{mtg_rate}
\title{Mortgage rate data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2571 rows and 3 columns.
}
\source{
\href{http://www.freddiemac.com/pmms/}{Freddie Mac PMMS}
}
\usage{
data(mtg_rate)
}
\description{
Time series of U.S. weekly average 30-year mortgage rates from April 2, 1971 through July 2, 2020.
}
\details{
Data are from the Freddie Mac Primary Mortgage Market Survey.

Rates are expressed in percentage points
3 columnns:
date: date of survey
rate: U.S. weekly average 30-year mortgage rate in percentage points
pts:  Average discount points and origination fees in percentage points
}
\examples{
data(mtg_rate)
}
\keyword{datasets}
