% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dark2.R
\name{theme_dark2}
\alias{theme_dark2}
\title{Alternative dark theme for ggplot2}
\usage{
theme_dark2(base_size = 12, base_family = "Courier New")
}
\arguments{
\item{base_size}{base font size, given in pts. defaults to 12}

\item{base_family}{base font family, defaults to Courier New}
}
\description{
A dark themed function
}
\examples{
ggplot(data=NULL,aes(x=1:5,y=1:5))+geom_line()+theme_dark2()
}
\keyword{theme}
